+++ apps/docs/src/content/02-foundations/04-ui-patterns/02-anlegeprozess/examples/project-illustrated-message.tsx
@@ -1,0 +1,27 @@
+import {
+  ActionGroup,
+  Button,
+  Heading,
+  IconProject,
+  IllustratedMessage,
+  Text,
+  LayoutCard,
+} from "@mittwald/flow-react-components";
+
+<LayoutCard>
+  <IllustratedMessage>
+    <IconProject />
+    <Heading>Erstes Projekt anlegen</Heading>
+    <Text>
+      Du hast noch keine Projekte angelegt! Lege ein Projekt
+      an und schon kannst du mit dem Entwickeln loslegen.
+    </Text>
+
+    <ActionGroup>
+      <Button color="secondary" variant="soft">
+        Tarif bestellen
+      </Button>
+      <Button color="accent">Anlegen</Button>
+    </ActionGroup>
+  </IllustratedMessage>
+</LayoutCard>;