+++ CHANGELOG.md
@@ -3,6 +3,83 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0-alpha.521](https://github.com/mittwald/flow/compare/0.2.0-alpha.520...0.2.0-alpha.521) (2025-10-17)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
+# [0.2.0-alpha.520](https://github.com/mittwald/flow/compare/0.2.0-alpha.519...0.2.0-alpha.520) (2025-10-17)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
+# [0.2.0-alpha.519](https://github.com/mittwald/flow/compare/0.2.0-alpha.518...0.2.0-alpha.519) (2025-10-17)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
+# [0.2.0-alpha.518](https://github.com/mittwald/flow/compare/0.2.0-alpha.517...0.2.0-alpha.518) (2025-10-16)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
+# [0.2.0-alpha.517](https://github.com/mittwald/flow/compare/0.2.0-alpha.516...0.2.0-alpha.517) (2025-10-15)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
+# [0.2.0-alpha.516](https://github.com/mittwald/flow/compare/0.2.0-alpha.515...0.2.0-alpha.516) (2025-10-10)
+
+### Bug Fixes
+
+* **Checkbox:** fix issue with RHF ([a8951e4](https://github.com/mittwald/flow/commit/a8951e49a7cae87abce0d5ea2e0522f8563b9599))
+* **PropsContext:** fix nesting level changes during re-render ([#2068](https://github.com/mittwald/flow/issues/2068)) ([31773e3](https://github.com/mittwald/flow/commit/31773e335bbb6b76a2d192d461c248c3a331b48b))
+
+# [0.2.0-alpha.515](https://github.com/mittwald/flow/compare/0.2.0-alpha.514...0.2.0-alpha.515) (2025-10-09)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
+# [0.2.0-alpha.514](https://github.com/mittwald/flow/compare/0.2.0-alpha.513...0.2.0-alpha.514) (2025-10-09)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
+# [0.2.0-alpha.513](https://github.com/mittwald/flow/compare/0.2.0-alpha.512...0.2.0-alpha.513) (2025-10-09)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
+# [0.2.0-alpha.512](https://github.com/mittwald/flow/compare/0.2.0-alpha.511...0.2.0-alpha.512) (2025-10-09)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
+# [0.2.0-alpha.511](https://github.com/mittwald/flow/compare/0.2.0-alpha.510...0.2.0-alpha.511) (2025-10-09)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
+# [0.2.0-alpha.510](https://github.com/mittwald/flow/compare/0.2.0-alpha.509...0.2.0-alpha.510) (2025-10-08)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
+# [0.2.0-alpha.509](https://github.com/mittwald/flow/compare/0.2.0-alpha.508...0.2.0-alpha.509) (2025-10-08)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
+# [0.2.0-alpha.508](https://github.com/mittwald/flow/compare/0.2.0-alpha.507...0.2.0-alpha.508) (2025-10-08)
+
+**Note:** Version bump only for package @mittwald/flow-project
+
+# [0.2.0-alpha.507](https://github.com/mittwald/flow/compare/0.2.0-alpha.506...0.2.0-alpha.507) (2025-10-07)
+
+### Bug Fixes
+
+* **Ext-Bridge:** fix validation issue in session token ([b288441](https://github.com/mittwald/flow/commit/b2884417048e0f1be1b22da48b4e2c14404f28d2))
+
+# [0.2.0-alpha.506](https://github.com/mittwald/flow/compare/0.2.0-alpha.505...0.2.0-alpha.506) (2025-10-07)
+
+### Bug Fixes
+
+* **Action:** fix confirmation modal not working ([09edaf9](https://github.com/mittwald/flow/commit/09edaf9a885446bf1a90db37611f29b154dc2f23))
+
+# [0.2.0-alpha.505](https://github.com/mittwald/flow/compare/0.2.0-alpha.504...0.2.0-alpha.505) (2025-10-07)
+
+### Features
+
+* **Icon:** add IconDnsSubZone ([#2058](https://github.com/mittwald/flow/issues/2058)) ([38e0198](https://github.com/mittwald/flow/commit/38e0198df9886dca9e961f604d3d881bd8c19415))
+
 # [0.2.0-alpha.504](https://github.com/mittwald/flow/compare/0.2.0-alpha.503...0.2.0-alpha.504) (2025-10-01)
 
 **Note:** Version bump only for package @mittwald/flow-project